guest user

+++ e_commerce/dennis_page/ecommerce/store/models.py
@@ -13,7 +13,7 @@ class Customer(models.Model):
 
 class Product(models.Model):
     name = models.CharField(max_length=200, null=True)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=7, decimal_places=2)
     digital = models.BooleanField(default=False, blank=False, null=True)
     image = models.ImageField(null=True, blank=True)
